Re-enable submit button if tokenization fails
Because we prevent the form from submitting if tokenization fails, we
need to re-enable the submit button. I discovered running this in a
sandbox that jquery-ujs will automatically disable the submit button
because the submit_tag handles adds `data-disable-with` to the button.
UJS disables the button in a setTimeout, which prevents us from
immediately re-enabling it. The only two options were to either stop
propagation, which could prevent users from adding other submit
handlers, or to re-enable it in a setTimeout that waits longer.