Pass order state to transaction import
Users might not always want the order to be advanced all the way through
to confirm. This allows a state parameter to be sent to the create
action on the transactions controller to specify the desired end state
for the order. If none is provided, the default value 'confirm' is used.